RDKCOM-4784: RDKBDEV-2642 fix incorrect change to support 64bit (#16)
Reason for change:
fix incorrect change to support 64bit targets

It's not valid to cast a pointer to a 32bit value to be a pointer to
a 64bit value.

Test Procedure: Sanity.
Risks: None.

+++ source/TR-181/middle_layer_src/wanmgr_dml_apis.c
@@ -189,7 +189,8 @@ BOOL WanManager_SetParamStringValue(ANSC_HANDLE hInsContext, char* ParamName, ch
         if (strcmp(ParamName, "Data") == 0)
         {
             char *webConf = NULL;
-            ULONG webSize = 0;
+            size_t webSize = 0;
+
             webConf = AnscBase64Decode(pString, &webSize);
             if(!webConf)
             {
@@ -212,7 +213,7 @@ BOOL WanManager_SetParamStringValue(ANSC_HANDLE hInsContext, char* ParamName, ch
         else if (strcmp(ParamName, "WanFailoverData") == 0)
         {
             char *webConf = NULL;
-            ULONG webSize = 0;
+            size_t webSize = 0;
 
             webConf = AnscBase64Decode(pString, &webSize);
             if(!webConf)
